Delete app.py
This code load balanced requests in a round robin way to docker containers and also did the scaling. It is deleted because it is project specific, i.e it works only for specific API's written for my university cloud computing course project. This file is replaced with a generic container orchestrator that will performs load balancing, auto-scaling and is fault-tolerant ( Performs health check on all containers and restarts faulty containers).